Refactor tr diffs and grouping into own modules
